feat: import first half of dacs using template
