--- packages/config/src/projects/other/da-beat/dac/bridges/hychain.ts
@@ -1,86 +1,0 @@
-import { ChainId } from '@l2beat/shared-pure'
-import { ProjectDiscovery } from '../../../../../discovery/ProjectDiscovery'
-import { hychain } from '../../../../layer2s/hychain'
-import {
-  DaAccessibilityRisk,
-  DaAttestationSecurityRisk,
-  DaExitWindowRisk,
-} from '../../types'
-import { DaBridge } from '../../types/DaBridge'
-import { DacTransactionDataType } from '../../types/DacTransactionDataType'
-import { toUsedInProject } from '../../utils/to-used-in-project'
-
-const discovery = new ProjectDiscovery('hychain')
-
-const DAC = discovery.getContractValue<{
-  membersCount: number
-  requiredSignatures: number
-}>('SequencerInbox', 'dacKeyset')
-const { membersCount, requiredSignatures } = DAC
-
-export const hychainDac = {
-  id: 'hychain-dac-bridge',
-  type: 'DAC',
-  display: {
-    name: 'Hychain DAC',
-    slug: 'dac',
-    description: 'Hychain DAC on Ethereum.',
-    links: {
-      websites: ['https://hychain.com'],
-      apps: ['https://bridge.hychain.com'],
-      documentation: ['https://docs.hychain.com'],
-      explorers: ['https://explorer.hychain.com'],
-      repositories: ['https://github.com/kintoxyz'],
-      socialMedia: [
-        'https://x.com/HYCHAIN_GAMES',
-        'https://discord.gg/hytopiagg',
-        'https://hychain.substack.com/',
-      ],
-    },
-  },
-  contracts: {
-    addresses: [
-      discovery.getContractDetails(
-        'SequencerInbox',
-        'Main entry point for the Sequencer submitting transaction batches.',
-      ),
-    ],
-    risks: [],
-  },
-  technology: `## Simple DA Bridge
-    The DA bridge is a smart contract verifying a data availability claim from DAC Members via signature verification.
-    The bridge requires a ${requiredSignatures}/${membersCount} threshold of signatures to be met before the data commitment is accepted.
-  `,
-  permissions: [
-    // Members: DAC uses BLS sigs, not EOAs
-    {
-      name: 'Sequencers',
-      accounts: discovery.getPermissionsByRole('Sequencer'),
-      description:
-        'Central actors allowed to submit transaction batches to the Sequencer Inbox.',
-      chain: discovery.chain,
-    },
-    {
-      name: 'RollupOwner',
-      accounts: discovery.getAccessControlRolePermission(
-        'UpgradeExecutor',
-        'EXECUTOR_ROLE',
-      ),
-      description:
-        'Multisig that can upgrade authorized batch posters via the UpgradeExecutor contract.',
-    },
-  ],
-  chain: ChainId.ETHEREUM,
-  requiredMembers: requiredSignatures,
-  totalMembers: membersCount,
-  transactionDataType: DacTransactionDataType.TransactionDataCompressed,
-  members: {
-    type: 'unknown',
-  },
-  usedIn: toUsedInProject([hychain]),
-  risks: {
-    attestations: DaAttestationSecurityRisk.NotVerified,
-    accessibility: DaAccessibilityRisk.NotEnshrined,
-    exitWindow: DaExitWindowRisk.Immutable,
-  },
-} satisfies DaBridge
