--- packages/config/src/projects/other/da-beat/dac/bridges/immutablex.ts
@@ -1,103 +1,0 @@
-import { ChainId } from '@l2beat/shared-pure'
-import { ProjectDiscovery } from '../../../../../discovery/ProjectDiscovery'
-import { getCommittee } from '../../../../../discovery/starkware'
-import { immutablex } from '../../../../layer2s/immutablex'
-import {
-  DaAccessibilityRisk,
-  DaAttestationSecurityRisk,
-  DaExitWindowRisk,
-} from '../../types'
-import { DaBridge } from '../../types/DaBridge'
-import { DacTransactionDataType } from '../../types/DacTransactionDataType'
-import { toUsedInProject } from '../../utils/to-used-in-project'
-
-const discovery = new ProjectDiscovery('immutablex')
-const committee = getCommittee(discovery)
-
-export const immutableXDac = {
-  id: 'immutablex-dac-bridge',
-  type: 'DAC',
-  display: {
-    name: 'ImmutableX DAC',
-    slug: 'dac',
-    description: 'ImmutableX DAC on Ethereum.',
-    links: {
-      websites: ['https://immutablex.xyz/'],
-      documentation: ['https://docs.immutablex.xyz/'],
-      repositories: ['https://github.com/Immutablex/immutablex'],
-      apps: ['https://app.immutable.com/'],
-      explorers: ['https://explorer.immutable.com/'],
-      socialMedia: ['https://twitter.com/Immutable'],
-    },
-  },
-  contracts: {
-    addresses: [
-      discovery.getContractDetails(
-        'Committee',
-        'Data Availability Committee (DAC) contract verifying data availability claim from DAC Members (via multisig check).',
-      ),
-    ],
-    risks: [],
-  },
-  technology: `## Simple DA Bridge
-    The DA bridge is a smart contract verifying a data availability claim from DAC Members via signature verification.
-    The bridge requires a ${committee.minSigners}/${committee.accounts.length} threshold of signatures to be met before the data commitment is accepted.
-  `,
-  permissions: [
-    {
-      name: 'Committee Members',
-      description: `List of addresses authorized to sign data commitments for the DA bridge.`,
-      accounts: committee.accounts.map((operator) => ({
-        address: operator.address,
-        type: 'EOA',
-      })),
-    },
-  ],
-  chain: ChainId.ETHEREUM,
-  requiredMembers: committee.minSigners,
-  totalMembers: committee.accounts.length,
-  transactionDataType: DacTransactionDataType.StateDiffs,
-  members: {
-    type: 'public',
-    list: [
-      {
-        name: 'Immutable',
-        href: 'https://assets.website-files.com/646557ee455c3e16e4a9bcb3/6499367de527dd82ab7475a3_Immutable%20Whitepaper%20Update%202023%20(3).pdf',
-      },
-      {
-        name: 'StarkWare',
-        href: 'https://assets.website-files.com/646557ee455c3e16e4a9bcb3/6499367de527dd82ab7475a3_Immutable%20Whitepaper%20Update%202023%20(3).pdf',
-      },
-      {
-        name: 'Deversifi',
-        href: 'https://assets.website-files.com/646557ee455c3e16e4a9bcb3/6499367de527dd82ab7475a3_Immutable%20Whitepaper%20Update%202023%20(3).pdf',
-      },
-      {
-        name: 'Consensys',
-        href: 'https://assets.website-files.com/646557ee455c3e16e4a9bcb3/6499367de527dd82ab7475a3_Immutable%20Whitepaper%20Update%202023%20(3).pdf',
-      },
-      {
-        name: 'Nethermind',
-        href: 'https://assets.website-files.com/646557ee455c3e16e4a9bcb3/6499367de527dd82ab7475a3_Immutable%20Whitepaper%20Update%202023%20(3).pdf',
-      },
-      {
-        name: 'Iqlusion',
-        href: 'https://assets.website-files.com/646557ee455c3e16e4a9bcb3/6499367de527dd82ab7475a3_Immutable%20Whitepaper%20Update%202023%20(3).pdf',
-      },
-      {
-        name: 'Infura',
-        href: 'https://assets.website-files.com/646557ee455c3e16e4a9bcb3/6499367de527dd82ab7475a3_Immutable%20Whitepaper%20Update%202023%20(3).pdf',
-      },
-      {
-        name: 'Cephalopod',
-        href: 'https://assets.website-files.com/646557ee455c3e16e4a9bcb3/6499367de527dd82ab7475a3_Immutable%20Whitepaper%20Update%202023%20(3).pdf',
-      },
-    ],
-  },
-  usedIn: toUsedInProject([immutablex]),
-  risks: {
-    attestations: DaAttestationSecurityRisk.SigVerified(true),
-    accessibility: DaAccessibilityRisk.NotEnshrined,
-    exitWindow: DaExitWindowRisk.Immutable,
-  },
-} satisfies DaBridge
